Updated setup scripts

+++ config/plab-client/setup-after-imaging.sh
@@ -1,13 +1,30 @@
 #!/bin/bash
 
-# First bless the rEFInd boot loader, otherwise Windows will start by default
-mkdir /Volumes/esp &&
-mount -t msdos /dev/disk0s1 /Volumes/esp &&
-bless --mount /Volumes/esp --setBoot --file /Volumes/esp/EFI/refind/refind_x64.efi --shortform &&
-umount /Volumes/esp &&
-echo "rEFInd blessed successfully"
+OS=$(uname -s)
 
-rmdir /Volumes/esp
+# Determine the name of the ethernet adapter
+case "$OS" in
+	"Linux" )
+		ETH_DEVICE="eth0" ;;
+	"Darwin" )
+		ETH_DEVICE="en0" ;;
+	* )
+		echo "Operating system $OS not supported" ;;
+esac
+
+
+if [[ "$OS" == "Darwin" ]];
+then
+	# First bless the rEFInd boot loader, 
+	# otherwise Windows will start by default
+
+	mkdir /Volumes/esp &&
+	mount -t msdos /dev/disk0s1 /Volumes/esp &&
+	bless --mount /Volumes/esp --setBoot --file /Volumes/esp/EFI/refind/refind_x64.efi --shortform &&
+	umount /Volumes/esp &&
+	echo "rEFInd blessed successfully"
+	rmdir /Volumes/esp
+fi
 
 
 # Defaults
@@ -15,10 +32,17 @@ hostname="Psychophysics-Unknown"
 trackpad=""
 
 # Fill parameters based on MAC address
-mac=$(ifconfig en0 | grep -o -E '([[:xdigit:]]{1,2}:){5}[[:xdigit:]]{1,2}')
-ip=$(ifconfig en0 | awk '/inet /{print $2}')
+mac=$(ifconfig $ETH_DEVICE | grep -o -E '([[:xdigit:]]{1,2}:){5}[[:xdigit:]]{1,2}')
 
-echo $mac
+if [[ "$OS" == "Darwin" ]];
+then
+	ip=$(ifconfig $ETH_DEVICE | awk '/inet /{print $2}')
+else
+	ip=$(ifconfig $ETH_DEVICE | grep -o -E '([[:digit:]]{1,3}.){3}[[:digit:]]{1,3}' | head -n 1)
+fi
+
+echo "Ethernet device: $ETH_DEVICE; MAC Address: $mac; IP Address: $ip"
+
 
 case "$mac" in
 	'ac:87:a3:22:e1:0d' )
@@ -47,13 +71,26 @@ case "$mac" in
 		;;
 esac
 
-# Set hostname
-scutil --set HostName $hostname &&
-echo "Host name set to $hostname"
 
-# Set trackpad ID
-plutil -replace BRPairedDevices -xml "<array><string>$trackpad</string></array>" /Library/Preferences/com.apple.Bluetooth.plist &&
-echo "Trackpad $trackpad setup"
+if [[ "$OS" == "Darwin" ]];
+then
+	# Set hostname
+	scutil --set HostName $hostname &&
+	echo "Hostname is set to $hostname"
+
+	# Set trackpad ID
+	plutil -replace BRPairedDevices -xml "<array><string>$trackpad</string></array>" /Library/Preferences/com.apple.Bluetooth.plist &&
+	echo "Trackpad $trackpad setup"
+fi
+
+
+if [[ "$OS" == "Linux" ]];
+then
+	hostname $hostname
+	echo $hostname >> /etc/hostname
+	echo "Hostname is set to $hostname"
+fi
+
 
 # Set name in DNS
 hostnamelc=$(echo $hostname | tr '[:upper:]' '[:lower:]')
